pyproject.toml
[tool.poetry]
name = "ddlc_api"
version = "0.1.0"
description = "An API with data of DDLC"
authors = ["Eliaz Bobadilla <eliaz.bobadilladev@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
pymongo = "^3.12.1"
python-dotenv = "^0.19.1"
dnspython = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^21.10b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
